fedora-install-intelliju: refactor and rename
There are a lot of changes here. I renamed the file and appended a "u" on the end to represent that this is installing IntelliJ Idea Ultimate, not the community edition. I also rebased my script on the Fedy installation script which has several benefits including the ability to download the latest version, downloading to the cache and preventing duplicate desktop launchers. This is also compatible with Fedy, so user can uninstall this through the GUI if they want to.

Credit: https://github.com/folkswithhats/fedy/blob/master/plugins/intellijideaultimate.plugin/install.sh